+++ trunk/snanalizer/java-src/snanalizer/services/UsuariosService.java
@@ -8,7 +8,7 @@ public interface UsuariosService {
 
 	public List<Usuario> getAll();
 	
-	public void modificar(Long id, String email, String password, String rol);
+	public void modificar(Integer id, String email, String password, String rol);
 	
 	public void crear(String email, String password, String rol);
